Add files via upload
A DTH provider is facing a lot of competition in the current market and it has become a challenge to retain the
existing customers in the current situation. Hence, the company wants to develop a model through which they
can do churn prediction of the accounts and provide segmented offers to potential churners. In this company,
account churn is a major issue because 1 account can have multiple customers. Hence by losing one account the
company might be losing more than one customer.
The current project is aimed at developing a churn prediction model for this company and to provide business
recommendations on a campaign focused on retaining customers. The campaign recommendation should be such
that it does not entail a huge cost for retention of customers and should remain within a budget earmarked for
this purpose.
1.2. Need for the project
A DTH provider’s biggest cost is cost of acquisition of a new customer. A customer thus acquired, will need to be
retained for quite a few years so that the initial cost of acquisition is recovered back and that particular account is
profitable1
. Due to this reason, customer churn directly impacts the profitability of a DTH operator. DTH providers
also are in a constant pressure to increase their customer base to maintain their profitability as most of them have
a fixed broadcaster/content provider fee irrespective of the number of customers in their customer base2. So,
more the number of customers, greater their profitability. Hence it becomes very important to not only increase
the customer base but also protect the current customer base.
Acquiring a new customer can cost five times more than retaining an existing customer. Increasing customer
retention by 5% can increase profits from 25-95%.3
As customer churn directly impacts both the top-line and bottom-line revenue of the business, existing customer
base needs to be protected. Providing all customers with offers to retain them would make a dent in the
profitability and hence it is very important to focus only on select set of customers who are at a higher risk of
churning.
1.3. Project Objectives
1) This project aims at identifying customers who have a propensity to churn so that targeted campaign
offers can be provided to them. This would ensure better top-line and bottom-line revenue for the
business.
2) Other insights from Exploratory data analysis and best performing model(s) will also be used to come up
with business recommendations.
8
2. Data report
2.1. Data collection
• The dataset contains ‘Account’ level data which is master data along with features of account such as
gender, marital status, city tier, account user count of primary account holder, whether the account is live
or churned, segment the account belongs to. It also possibly contains certain derived features such as
tenure (which probably could have been derived from account open date)
• The dataset also contains information taken/derived from transaction data and rolled up at Account level
and given as a feature for the account – for e.g., Number of days since none of the account holders
contacted customer care, monthly average cashback for last 12 months, number of complaints made last
year, number of times customer care contacted last year, revenue per month in last 12 months, how many
times customers have used coupons to pay in last 12 months, satisfaction score and customer service
score.
• Most of the transaction data roll-up at account level has been done for 12 months (previous year).
However, the revenue growth percentage has been taken for last year in comparison with previous one
year which implies that 24-months’ worth of data has been used to calculate this field.
• As the dataset has been provided, the methodology used by customer to extract data is not known. The
frequency at which this dataset is extracted has also not been specified.